Second Iteraction
Added the pre-processing file.
Colors changed to a white to red gradient, and the background changed
to a light blue to make the white stand out.
The `dy` and `dx` text attributes tweaked to address the label
alignment issue.
An extra ten pixels were added to the bottom margin to give some space
between the tick and axis labels.