Final Iteration
The pre-processing file was changed to add the airport name to the data
using data found here: https://github.com/jbrooksuk/JSON-Airports.
The graph axes were switched around as per KU’s suggestion.
The graph was further amended by moving the x-axis to the top, and
nudging the axis labels towards the top-left corner.
The tooltip now gives the airport name and all the data – an extra key
was added to the `d3.nest()` data to facilitate this.