Fifth Iteration
The detail about the number of flights per airport. Flights per airport
are still used to filter the data, but this detail is no longer
expressed in the chart.
Blue gradient added to the bars in the absence of the number of flight
coloring.
The tooltip now disappears on mouse out, and the opacity of the bars
also varies according to the position of the mouse.